+++ src/Webhook/Handler/AuthorizationVoided.php
@@ -7,6 +7,7 @@
 
 namespace Swag\PayPal\Webhook\Handler;
 
+use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStates;
 use Shopware\Core\Framework\Context;
 use Swag\PayPal\RestApi\PayPalApiStruct;
 use Swag\PayPal\RestApi\V1\Api\Webhook as WebhookV1;
@@ -38,6 +39,8 @@ class AuthorizationVoided extends AbstractWebhookHandler
             $orderTransaction = $this->getOrderTransactionV2($authorization, $context);
         }
 
-        $this->orderTransactionStateHandler->cancel($orderTransaction->getId(), $context);
+        if ($this->isChangeAllowed($orderTransaction, OrderTransactionStates::STATE_CANCELLED)) {
+            $this->orderTransactionStateHandler->cancel($orderTransaction->getId(), $context);
+        }
     }
 }