+++ tests/Webhook/Handler/AuthorizationVoidedTest.php
@@ -38,6 +38,12 @@ class AuthorizationVoidedTest extends AbstractWebhookHandlerTestCase
         $this->assertInvokeWithoutTransaction(WebhookEventTypes::PAYMENT_AUTHORIZATION_VOIDED, $webhook, $reason);
     }
 
+    public function testInvokeWithSameInitialState(): void
+    {
+        $webhook = $this->createWebhookV1();
+        $this->assertInvoke(OrderTransactionStates::STATE_CANCELLED, $webhook, OrderTransactionStates::STATE_CANCELLED);
+    }
+
     protected function createWebhookHandler(): AuthorizationVoided
     {
         return new AuthorizationVoided(